Backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pymongo==3.12.3
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3
drf-yasg==1.21.7
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
pytz==2023.3
setuptools
django-extensions
sqlparse==0.2.4
dnspython